% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{api}
\alias{api}
\title{Antecedent Precipitation Index}
\usage{
api(precipitation, k, t)
}
\arguments{
\item{precipitation}{A vector of precipitation estimates/measurements}

\item{k}{Decay constant}

\item{t}{Number of days to account for, usually 5/7/14}
}
\description{
API
}
