% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/degreeday.R
\name{degree_day}
\alias{degree_day}
\title{The degree-day model}
\usage{
degree_day(Ta, Tt, P, C_WH, init_ts = 0)
}
\arguments{
\item{Ta}{}

\item{Tt}{}

\item{P}{}

\item{C_WH}{}
}
\description{
A simplem degree day model. Whenever the observed air temperatture (T) celsius is 
lower than a threshold temperature Tt, precipitation is added to the snow storage
(accumulation A [mm]). In addition to the accumulation, the liquid water sontent 
S_lq (mm) in the snowpack is also calculated. S_lq is calculated accounting
for precipitation P (mmm), melt M (mm) and refreezing (R) (mm) and has an upper
bound constrained by the water holding capacity `C_WH` (-). Refreezing is detemined
by S_lq of the day before, a degree-day factor C_M (mm/day C) and a refreezing
factor C_FR (-). Melt is constrined by the preceding accumulation and calculated 
using C_M, Tt and T. The contribution to surface runoff Q (mm) is all water that
exceeds C_WH of the snow pack. For this study, C_WH was kept constant at a value
0.1.
}
