% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.KGE2011cpp}
\alias{.KGE2011cpp}
\title{Kling-Gupta Efficiency}
\usage{
.KGE2011cpp(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector of same length as x}
}
\value{
KGE value
}
\description{
Kling-Gupta efficiency function
}
