% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.LMEcpp}
\alias{.LMEcpp}
\title{Liu Mean Efficiency}
\usage{
.LMEcpp(x, y)
}
\arguments{
\item{x}{numeric vector of observed values}

\item{y}{numeric vector of simulated values}
}
\value{
LME value
}
\description{
Liu Mean Efficiency wrapper function. This criteria aims to \deqn{\text{LME} = 1 - \sqrt{(k_1 - 1)^2 + (\beta - 1)^2}}
}
