% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{.NSEcpp}
\alias{.NSEcpp}
\title{Nash-Sutcliffe Efficiency}
\usage{
.NSEcpp(x, y)
}
\arguments{
\item{x}{numeric vector of observed values}

\item{y}{numeric vector of simulated values}
}
\value{
KGE value
}
\description{
Nash-Sutcliffe Efficiency wrapper function
}
