% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{kge}
\alias{kge}
\title{Kling-Gupta Efficiency}
\usage{
kge(x, y)
}
\arguments{
\item{x}{numeric vector}

\item{y}{numeric vector of same length as x}
}
\value{
KGE value
}
\description{
Kling-Gupta efficiency function
}
