% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crits.R
\name{NSE}
\alias{NSE}
\title{Nash-Sutcliffe criteria wrapper}
\usage{
NSE(x, y)
}
\arguments{
\item{x}{Observed series}

\item{y}{Simulated series}

\item{type}{Version of computation}
}
\description{
The Nash-Sutcliffe criteria. The wrapper takes care of the errors
as well as the argument order in the call
}
