% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{triangle_vertical_end_colourbar}
\alias{triangle_vertical_end_colourbar}
\title{colorbar triangular ends for colorbar}
\usage{
triangle_vertical_end_colourbar(...)
}
\arguments{
\item{...}{}
}
\description{
ggplot2 compatible triangular ending for colorbar such as is in ProPlot.
}
